src/problem.py
'''
Created on Aug 18, 2016

@author: lenovo
'''

import operator #TODO: What does this do?
import sys

#class State(object): #TODO: Why an object? I want a function to be able to take a human-readable representation of the game state and return its value
#    '''
#    '''
    
def __init__(self, simstate): #TODO: Need a list of goals (high-level actions?)
    '''
    Takes a visual representation of the current state and parses it into an internal representation. (Yes, I've been thinking about knowledge representation lately) 
    '''
    
    
def parse(simstate): #Could recursively split first and rest and send rest to the parse function. Function returns a list of units and their coordinates.
    '''
    '''  
    state = {}   
    y=0
    x=0    
    for cell in simstate:   
        if cell in 'HB*!':
            state[(x,y)]=cell       
        elif cell=='\n':
            y += 1
            x=-1 #Hmm...        
        x += 1    
    return state

def write(state, action=None):
    '''
    A state is a dictionary of positions and objects, keyed on position.
    '''
    ordered = sorted(state.items(), key=operator.itemgetter(0)) #TODO: 
    simstate = ''
    cell = (-1, 0)
    for object in ordered:
        location = object[0]
        name = object[1] #TODO: This name isn't quite accurate since objects could be obstacles 
        simstate += '\n' * (location[1] - cell[1])
        y = location[1]
        simstate += ' ' * (location[0] - cell[0] - 1)
        simstate += name
        x = location[0]
    simstate += '\n'
    return simstate

def c(state): #TODO: I really want to make this a higher-order function
    return 0 #TODO: For now you get nothing special for any particular configuration of the world

def applicable_actions(state): 
    '''
    Returns an iterable list of actions applicable in the given state.
    '''
    return [1]

def transition(state, action):
    '''
    Returns the next state (s') from the current state (s) given an action.
    '''
    return state + action
        
if __name__ == '__main__': #Read in a simulated state and write it out
    pass
    